+++ pytsite/sitemap/__init__.py
@@ -5,4 +5,5 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
+# Public API
 from ._sitemap import Sitemap, Index